refactor search.js compoent and separate children components
